pyproject.toml
[project]
name = "bnet_simulator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "CesareDev", email = "corsicesare.lavoro@gmail.com" },
    { name = "LuPierba", email = "filo.pierbatti3@gmail.com" }
]
requires-python = ">=3.11"
dependencies = [
    "matplotlib>=3.10.3",
    "pandas>=2.2.3",
    "pygame>=2.6.1",
    "pyyaml>=6.0.3",
    "scipy>=1.15.3",
    "tqdm>=4.67.1",
]

[project.scripts]
sim = "run:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.hatch.build]
sources = ["src"]
